refactor: migrate citation imports to citation-service (#1479 Sub-PR C)
Migrate ~20 files from direct citation-archive/citation-auditor imports
to the unified citation-service.ts re-exports. This is a pure import
path migration with no behavior changes.

Migrated to citation-service.ts:
- extractCitationsFromContent (7 consumers)
- auditCitations + types (5 consumers)
- MIN_SOURCE_CONTENT_LENGTH (2 consumers)
- AuditResult, CitationAudit, AuditRequest types (4 consumers)

Kept on citation-archive.ts (not re-exported by service):
- verifyCitationsForPage (writes YAML archive — needed by CI pipeline)
- readCitationArchive, listArchivedPages (YAML archive operations)
- extractClaimSentence (granular helper, 1 consumer)

Cleanup:
- Made fetchCitationUrl and FetchResult non-exported in citation-archive
  (only used internally by verifyCitationsForPage)
- Added auditCitations to citation-service.ts re-exports
