+++ crux/authoring/orchestrator/tools/audit-citations.ts
@@ -5,7 +5,7 @@
  * Cost: ~$0.20.
  */
 
-import { auditCitations, type AuditRequest } from '../../../lib/citation/citation-auditor.ts';
+import { auditCitations, type AuditRequest } from '../../../lib/citation/citation-service.ts';
 import type { ToolRegistration } from './types.ts';
 
 export const tool: ToolRegistration = {
