+++ crux/authoring/page-improver/phases/citation-audit.ts
@@ -18,7 +18,7 @@
  * See issue #670.
  */
 
-import { auditCitations, MIN_SOURCE_CONTENT_LENGTH, type AuditResult, type SourceCache } from '../../../lib/citation/citation-auditor.ts';
+import { auditCitations, MIN_SOURCE_CONTENT_LENGTH, type AuditResult, type SourceCache } from '../../../lib/citation/citation-service.ts';
 import type { SourceCacheEntry } from '../../../lib/content/section-writer.ts';
 import type { FetchedSource } from '../../../lib/search/source-fetcher.ts';
 import type { PageData, ResearchResult, PipelineOptions } from '../types.ts';
