+++ crux/authoring/page-improver/phases/research.ts
@@ -9,7 +9,7 @@
 import { z } from 'zod';
 import { MODELS } from '../../../lib/anthropic.ts';
 import { fetchSources, type FetchRequest, type FetchedSource } from '../../../lib/search/source-fetcher.ts';
-import { MIN_SOURCE_CONTENT_LENGTH } from '../../../lib/citation/citation-auditor.ts';
+import { MIN_SOURCE_CONTENT_LENGTH } from '../../../lib/citation/citation-service.ts';
 import type { SourceCacheEntry } from '../../../lib/content/section-writer.ts';
 import type { PageData, AnalysisResult, ResearchResult, PipelineOptions } from '../types.ts';
 import { log, writeTemp } from '../utils.ts';
